TfvcWorkspaceLocator: migrate getPartialWorkspace from CommandUtils
This method will be implemented in a general way on top of TfvcClient.
